+++ hbw/app/javascript/packs/hbw/components/helpers/index.js
@@ -1,12 +1,12 @@
 import withCallbacks from './callbacks';
-import withDeleteIf from './delete_if';
+import withConditions from './conditions';
 import withTasks from './tasks';
 import withSelect from './select';
 import { compose } from './utils';
 
 export {
   withCallbacks,
-  withDeleteIf,
+  withConditions,
   withTasks,
   withSelect,
   compose