+++ CHANGELOG.md
@@ -18,6 +18,7 @@ v1.6.0 [unreleased]
 -   [#218](https://github.com/latera/homs/pull/218) Pass initial BP variables to widget initial payload.
 -   [#223](https://github.com/latera/homs/pull/223) Add directories for postgresql-client.
 -   [#227](https://github.com/latera/homs/pull/227) Add `required` option for string inputs. Use `nullable: false` in select/select_table for the same reason.
+-   [#230](https://github.com/latera/homs/pull/230) Add disable_if condition to checkbox, datetime, file_list, select, select_table, string, text.
 
 ### Bugfixes
 -   [#194](https://github.com/latera/homs/pull/194) Set limit on the number of choices in lookup select from sql requests.